# openapi.yaml
openapi: 3.0.0
info:
  title: 'Star Citizen API'
  contact:
    email: foxftw@star-citizen.wiki
  version: 3.0.0
servers:
  -
    url: 'https://api.star-citizen.wiki'
paths:
...
  /api/comm-link-images:
    get:
      tags:
        - Comm-Links
        - RSI-Website
        - Images
      summary: 'Comm-Link Images'
      description: 'List available comm-link images with pagination.'
      operationId: a705f668107b86178c6624b972dddefd
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/page_number'
        -
          $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: 'List of Comm-Link Images'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/comm_link_image'
  /api/comm-link-images/random:
    get:
      tags:
        - Comm-Links
        - RSI-Website
        - Images
      summary: 'Comm-Link Images Random'
      description: 'Retrieve random comm-link images, optionally filtered by tag.'
      operationId: 0b0b5e244ccfb0fe78f81d536ba99257
      parameters:
        -
          name: limit
          in: query
          schema:
            type: integer
            maximum: 100
        -
          name: 'filter[tags]'
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 'Retrieve a random Comm-Link Image. Limit parameter sets the number of random images'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/comm_link_image'
  /api/comm-link-images/search:
    post:
      tags:
        - Comm-Links
        - RSI-Website
        - Images
        - Search
      summary: 'Comm-Link Image Search by filename'
      description: 'Search comm-link images by filename with optional tag filtering.'
      operationId: dbd72b035105c4dcf17ee654931fe2ff
      parameters:
        -
          name: 'filter[tags]'
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 'Search for a Comm-Link Image by its filename.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/comm_link_image'
  /api/galactapedia:
    get:
      tags:
        - Galactapedia
        - RSI-Website
      summary: 'Galactapedia Overview'
      description: 'Return paginated Galactapedia articles with category, tag, and template filters.'
      operationId: 48f80f5118a6ef2670845b9dd1b90357
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/page_number'
        -
          $ref: '#/components/parameters/page_size'
        -
          name: 'filter[category]'
          in: query
          schema:
            type: string
        -
          name: 'filter[categoryId]'
          in: query
          schema:
            type: string
        -
          name: 'filter[tag]'
          in: query
          schema:
            type: string
        -
          name: 'filter[tagId]'
          in: query
          schema:
            type: string
        -
          name: 'filter[template]'
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 'List of Galactapedia Articles'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/galactapedia_article'
...
  '/api/galactapedia/{id}':
    get:
      tags:
        - Galactapedia
        - RSI-Website
      summary: 'Galactapedia Article'
      description: 'Retrieve a Galactapedia article by ID with available includes and translations.'
      operationId: 541ad0798eef38797ca9a341b81eedac
      parameters:
        -
          $ref: '#/components/parameters/locale'
        -
          name: id
          in: path
          required: true
          schema:
            description: 'Galactapedia Article ID'
            type: string
        -
          name: include
          in: query
          explode: false
          allowReserved: true
          schema:
            description: 'Available Galactapedia includes'
            type: array
            items:
              type: string
              enum:
                - translations
                - tags
                - categories
                - related_articles
                - properties
      responses:
        '200':
          description: 'A singular Article'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/galactapedia_article'
        '404':
          description: 'No Article with specified ID found.'
  /api/galactapedia/search:
    post:
      tags:
        - Galactapedia
        - RSI-Website
        - Search
      summary: 'Galactapedia Article Search'
      description: 'Search Galactapedia articles by title, template, slug, or related metadata.'
      operationId: ed9b1b052b5dd9a03e0f7ee5a21cb8dd
      requestBody:
        description: 'Article (partial) title, template or slug'
        required: true
        content:
          application/json:
            schema:
              type: object
            example: '{"query": "Banu"}'
      responses:
        '200':
          description: 'List of articles matching the query'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/galactapedia_article'
        '404':
          description: 'No Article found.'
...
components:
  schemas:
...
    comm_link_image:
      title: 'Comm-Link Image'
      description: 'Image used in a Comm-Link'
      properties:
        rsi_url:
          type: string
        api_url:
          type: string
          nullable: true
        alt:
          type: string
        size:
          type: integer
        mime_type:
          type: string
        last_modified:
          type: string
        tags:
          type: array
          items:
            type: string
          nullable: true
        similar_url:
          type: string
      type: object
    galactapedia_article:
      title: 'Galactapedia Article'
      description: 'An article form the Galactapedia'
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        thumbnail:
          type: string
        template:
          type: string
          nullable: true
        rsi_url:
          type: string
        api_url:
          type: string
        created_at:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/galactapedia_category'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/galactapedia_tag'
        properties:
          type: array
          items:
            $ref: '#/components/schemas/galactapedia_property'
        related_articles:
          type: array
          items:
            $ref: '#/components/schemas/galactpedia_related_article'
        translations:
          oneOf:
            -
              type: string
            -
              type: array
              items:
                $ref: '#/components/schemas/translation'
      type: object
    galactapedia_category:
      title: 'Galctapedia article category'
      description: 'Category of an article'
      properties:
        id:
          type: string
        name:
          type: string
      type: object
    galactapedia_property:
      title: 'Galctapedia article property'
      description: 'Property of an article'
      properties:
        name:
          type: string
        value:
          type: string
      type: object
    galactpedia_related_article:
      title: 'Galactapedia related article'
      description: 'Related article for this galactapedia article'
      properties:
        id:
          type: string
        title:
          type: string
        url:
          type: string
        api_url:
          type: string
      type: object
    galactapedia_tag:
      title: 'Galctapedia article tag'
      description: 'Tag of an article'
      properties:
        id:
          type: string
        name:
          type: string
      type: object
...
    translation:
      title: 'Grouped Translations'
      description: 'Translations of an entity'
      properties:
        en:
          type: string
        de:
          type: string
        zh:
          type: string
      type: object
...
  parameters:
...
    page:
      name: page
      in: query
      description: 'Page number for pagination (starts at 1).'
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    page_number:
      name: 'page[number]'
      in: query
      description: 'Page number for pagination (starts at 1).'
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    page_size:
      name: 'page[size]'
      in: query
      description: 'Number of results to return per page.'
      required: false
      schema:
        type: integer
        default: 30
        maximum: 200
        minimum: 1
    locale:
      name: locale
      in: query
      description: 'Locale code for translated fields (e.g. en, de, zh).'
      required: false
      schema:
        type: string
...